Lambda-FastApi-RAG/requirements.txt
fastapi
uvicorn
boto3
python-dotenv
langchain_core
langchain
langchain_community
langchain_openai
